refactor: fix long method/god component code smell
Split large components into smaller, focused modules for better maintainability:

About Page (194 → 62 lines):
- Extract HeroSection component for profile info and social links
- Extract SkillsSection component for technical expertise display
- Extract ExperienceSection component with improved line rendering
- Extract AboutStructuredData component for centralized structured data

NavBar (153 → 84 lines):
- Extract MobileMenu component for mobile navigation
- Create useNavigation hook for navigation state management
- Create useMobileMenu hook for mobile menu logic and keyboard handling
- Fix type safety issues (remove 'as any' assertions)

Benefits:
- Single Responsibility Principle adherence
- Improved testability and reusability
- Better code organization and maintainability
- Enhanced type safety

All tests passing (118/118) with maintained coverage (84.98%)
